Reorganized a few test functions
