+++ src/test/test_state_manager.rs
@@ -1,5 +1,6 @@
 use core::panic;
 
+use bevy::input::mouse::MouseMotion;
 use bevy::prelude::*;
 use crate::juice_renderer::{self, draw_selection_circle};
 use crate::simulation::sim_state_manager::{
@@ -8,7 +9,7 @@ use crate::simulation::sim_state_manager::{
     add_faucet,
     add_drain,
 };
-use crate::simulation::SimSurfaceDirection;
+use crate::simulation::{step_simulation_once, SimSurfaceDirection};
 use crate::simulation::{
     self,
 	SimConstraints,
@@ -22,13 +23,13 @@ use crate::simulation::{
 		add_particle,
 	},
 };
-use crate::util::get_cursor_position;
+use crate::util::{cartesian_to_polar, get_cursor_position, polar_to_cartesian};
 
 /// Create a simulation layout for testing.
 pub fn construct_test_simulation_layout(
 	constraints:	&mut SimConstraints,
 	grid:			&mut SimGrid,
-	mut commands:	Commands) {
+	mut commands:	&mut Commands) {
 
 	// Create a lil cup
 	// grid.set_grid_cell_type(40, 32, SimGridCellType::Solid);
@@ -75,7 +76,7 @@ pub fn construct_test_simulation_layout(
     let drain_pos = Vec2::new(grid.cell_size as f32 * 25.0, 0.0);
     let surface_direction = None;
 
-    add_drain(&mut commands, grid, drain_pos, surface_direction).ok();
+    add_drain(commands, grid, drain_pos, surface_direction).ok();
 
 	// Spawn a small test group of particles at the center of the screen.
 	let grid_center: Vec2 = Vec2 {
@@ -92,7 +93,7 @@ pub fn construct_test_simulation_layout(
 	// );
 
 	let _moar_test_particles = add_particles_in_radius(
-		&mut commands,
+		commands,
         constraints,
 		grid,
 		1.75,
@@ -102,6 +103,124 @@ pub fn construct_test_simulation_layout(
 	);
 
 	println!("Creating a test simulation with {} particles...", constraints.particle_count);
+}
+
+/// Debugging state controller.
+pub fn debug_state_controller(
+	mut commands:		Commands,
+	keys:				Res<Input<KeyCode>>,
+	mouse:				Res<Input<MouseButton>>,
+	mut mouse_motion:	EventReader<MouseMotion>,
+	windows:			Query<&Window>,
+	cameras:			Query<(&Camera, &GlobalTransform)>,
+	mut constraints:	ResMut<SimConstraints>,
+	mut grid:			ResMut<SimGrid>,
+	mut particles:		Query<(Entity, &mut SimParticle)>) {
+
+	// Reset simulation when we press R.
+	if keys.just_pressed(KeyCode::R) {
+		crate::simulation::reset_simulation_to_default(
+			&mut commands,
+			constraints.as_mut(),
+			grid.as_mut(),
+			&mut particles
+		);
+		construct_test_simulation_layout(
+			constraints.as_mut(),
+			grid.as_mut(),
+			&mut commands
+		);
+		return;
+	}
+
+	// Rotate/scale gravity when we press the arrow keys.
+	let gravity_rotation: i8	= keys.pressed(KeyCode::Right) as i8 -
+									keys.pressed(KeyCode::Left) as i8;
+	let gravity_magnitude: i8	= keys.pressed(KeyCode::Up) as i8 -
+									keys.pressed(KeyCode::Down) as i8;
+	let mut polar_gravity: Vec2	= cartesian_to_polar(constraints.gravity);
+	polar_gravity.x				+= 200.0 * gravity_magnitude as f32 * constraints.timestep;
+	polar_gravity.y				+= 4.0 * gravity_rotation as f32 * constraints.timestep;
+
+	// Limit the magnitude of the vector to prevent ugly behavior near 0.0.
+	polar_gravity.x				= f32::max(0.0, polar_gravity.x);
+	constraints.gravity			= polar_to_cartesian(polar_gravity);
+
+	// Place/remove grid cells if the mouse is clicked on a cell.
+	let should_place_cell: bool		= mouse.pressed(MouseButton::Left);
+	let should_remove_cell: bool	= mouse.pressed(MouseButton::Right);
+
+	// Get the mouse's motion between this and the last frame.
+	let mut cursor_delta: Vec2 = Vec2::ZERO;
+	for event in mouse_motion.read() {
+		cursor_delta.x = event.delta.x;
+		cursor_delta.y = event.delta.y;
+	}
+
+	if should_place_cell {
+		let cursor_position: Vec2	= get_cursor_position(&windows, &cameras);
+		let cell_coordinates: Vec2	= grid.get_cell_coordinates_from_position(&cursor_position);
+		let _ = grid.set_grid_cell_type(
+			cell_coordinates.x as usize,
+			cell_coordinates.y as usize,
+			SimGridCellType::Solid
+		);
+
+		// Delete all particles in the cell we are turning into a solid.
+		let lookup_index: usize = grid.get_lookup_index(cell_coordinates);
+		grid.delete_all_particles_in_cell(
+			&mut commands,
+			constraints.as_mut(),
+			&particles,
+			lookup_index
+		);
+
+	} else if should_remove_cell {
+		let cursor_position: Vec2	= get_cursor_position(&windows, &cameras);
+		let cell_coordinates: Vec2	= grid.get_cell_coordinates_from_position(&cursor_position);
+		let _ = grid.set_grid_cell_type(
+			cell_coordinates.x as usize,
+			cell_coordinates.y as usize,
+			SimGridCellType::Air
+		);
+	}
+}
+
+/// Simulation state manager initialization.
+pub fn test_setup(
+	mut commands:		Commands,
+	mut constraints:	ResMut<SimConstraints>,
+	mut grid:			ResMut<SimGrid>) {
+
+	construct_test_simulation_layout(
+		constraints.as_mut(),
+		grid.as_mut(),
+		&mut commands
+	);
+
+}
+
+pub fn test_update(
+	mut constraints:	ResMut<SimConstraints>,
+	mut grid:			ResMut<SimGrid>,
+	mut particles:		Query<(Entity, &mut SimParticle)>,
+    faucets:		    Query<(Entity, &SimFaucet)>,
+    drains:		        Query<(Entity, &SimDrain)>,
+	commands:	        Commands,
+    ) {
+
+	// let delta_time: f32 = time.delta().as_millis() as f32 * 0.001;
+	let fixed_timestep: f32 = constraints.timestep;
+
+    step_simulation_once(
+        commands,
+        constraints.as_mut(),
+        grid.as_mut(),
+        &mut particles,
+        &faucets,
+        &drains,
+        fixed_timestep
+    );
 }
 
 /// Test particle selection.
@@ -214,8 +333,8 @@ fn add_faucet_test() {
     juicebox_test.insert_resource(SimConstraints::default());
 
     // Add our test setup environment
-	juicebox_test.add_systems(Startup, simulation::test_setup);
-	juicebox_test.add_systems(Update, simulation::test_update);
+	juicebox_test.add_systems(Startup, test_setup);
+	juicebox_test.add_systems(Update, test_update);
 
     // Add the test function for our add_faucet state change
     juicebox_test.add_systems(Update, test_add_faucet_update);
@@ -239,8 +358,8 @@ fn run_faucet_test() {
     juicebox_test.insert_resource(SimGrid::default());
     juicebox_test.insert_resource(SimConstraints::default());
 
-	juicebox_test.add_systems(Startup, simulation::test_setup);
-	juicebox_test.add_systems(Update, simulation::test_update);
+	juicebox_test.add_systems(Startup, test_setup);
+	juicebox_test.add_systems(Update, test_update);
 
     // Add the test function for our add_faucet state change
     juicebox_test.add_systems(Update, test_add_faucet_update);
@@ -272,8 +391,8 @@ fn add_drain_test() {
     juicebox_test.insert_resource(SimConstraints::default());
 
     // Add our test setup environment
-	juicebox_test.add_systems(Startup, simulation::test_setup);
-	juicebox_test.add_systems(Update, simulation::test_update);
+	juicebox_test.add_systems(Startup, test_setup);
+	juicebox_test.add_systems(Update, test_update);
 
     // Add the test function for our add_drain state change
     juicebox_test.add_systems(Update, test_add_drain_update);
@@ -297,8 +416,8 @@ fn drain_drain_test() {
     juicebox_test.insert_resource(SimGrid::default());
     juicebox_test.insert_resource(SimConstraints::default());
 
-	juicebox_test.add_systems(Startup, simulation::test_setup);
-	juicebox_test.add_systems(Update, simulation::test_update);
+	juicebox_test.add_systems(Startup, test_setup);
+	juicebox_test.add_systems(Update, test_update);
 
     // Add the test function for our add_drain state change
     juicebox_test.add_systems(Update, test_add_drain_update);
